% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-Report.R
\name{add_Groups_metadata}
\alias{add_Groups_metadata}
\title{Add group meta data for report}
\usage{
add_Groups_metadata(
  dfResults,
  dfGroups,
  strGroupLevel = c("Site", "Study", "Country"),
  strGroupDetailsParams
)
}
\arguments{
\item{dfResults}{`data.frame` Analysis results data.}

\item{dfGroups}{`data.frame` Analysis groups data.}

\item{strGroupLevel}{`string` denoting group level}

\item{strGroupDetailsParams}{`string`}
}
\value{
A modified `dfResults` which metadata attached
}
\description{
Add group meta data for report
}
