% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_Metric.R
\name{ChartsMetricDefault}
\alias{ChartsMetricDefault}
\title{Charts Metric Default Function}
\usage{
ChartsMetricDefault(
  lCharts,
  dfResults,
  lMetric,
  dfGroups,
  dfBounds,
  vThreshold,
  bDebug
)
}
\arguments{
\item{lCharts}{`list`}

\item{dfResults}{`data.frame` A stacked summary of analysis pipeline output.
Created by passing a list of results returned by [Summarize()] to
[BindResults()]. Expected columns: `GroupID`, `GroupLevel`, `Numerator`,
`Denominator`, `Metric`, `Score`, `Flag`, `MetricID`, `StudyID`,
`SnapshotDate`.}

\item{lMetric}{`list` Metric-specific metadata for use in charts and
reporting. Created by passing an `lWorkflow` object to [MakeMetric()] and
turing it into a list. Expected columns: `File`,`MetricID`, `Group`,
`Abbreviation`, `Metric`, `Numerator`, `Denominator`, `Model`, `Score`, and
`strThreshold`. For more details see the Data Model vignette:
`vignette("DataModel", package = "gsm.kri")`.}

\item{dfGroups}{`data.frame` Group-level metadata dictionary. Created by
passing CTMS site and study data to [MakeLongMeta()]. Expected columns:
`GroupID`, `GroupLevel`, `Param`, `Value`.}

\item{dfBounds}{`data.frame` Set of predicted percentages/rates and upper-
and lower-bounds across the full range of sample sizes/total exposure
values for reporting. Created by passing `dfResults` and `dfMetrics` to
[MakeBounds()]. Expected columns: `Threshold`, `Denominator`, `Numerator`,
`Metric`, `MetricID`, `StudyID`, `SnapshotDate`.}

\item{vThreshold}{`numeric` Threshold values for metric.}

\item{bDebug}{`logical` Print debug messages? Default: `FALSE`.}
}
\value{
A list containing the following charts:
- scatterPlot: A scatter plot using JavaScript.
- barChart: A bar chart using JavaScript with metric on the y-axis.
- metricTable: Table using JavaScript with metric scores.
}
\description{
`r lifecycle::badge("stable")`

The function creates all metric charts for a default metric using the data provided
}
\examples{
lCharts <- list()
strMetricID = "Analysis_kri0001"

dfResults_latest <-  gsm.core::reportingResults \%>\%
 dplyr::filter(MetricID == strMetricID) \%>\%
 FilterByLatestSnapshotDate()

dfBounds_latest <- gsm.core::reportingBounds \%>\%
 dplyr::filter(MetricID == strMetricID) \%>\%
 FilterByLatestSnapshotDate()

dfGroups <- gsm.core::reportingGroups

lMetric <- gsm.core::reportingMetrics \%>\%
 dplyr::filter(MetricID == strMetricID) \%>\%
 as.list()

vThreshold = gsm.core::ParseThreshold(lMetric$Threshold, bSort = FALSE)

lCharts <- lCharts \%>\%
  ChartsMetricDefault(
    dfResults= dfResults_latest,
    lMetric = lMetric,
    dfGroups = dfGroups,
    dfBounds = dfBounds_latest,
    vThreshold = vThreshold,
    bDebug = FALSE
  )

lCharts
}
