% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTQ_lineplot.R
\name{CTQ_lineplot}
\alias{CTQ_lineplot}
\title{CTQ Time Series Lineplot}
\usage{
CTQ_lineplot(dfResults, strCTQ)
}
\arguments{
\item{dfResults}{A results `data.frame` from the output of `gsm.reporting::BindResults()` used
to create a variety of visualizations like the line plot, bar plot.}

\item{strCTQ}{A `string` to label the QTL being measured}
}
\value{
A `plotly` object
}
\description{
CTQ Time Series Lineplot
}
