% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-Report.R
\name{FilterByFlags}
\alias{FilterByFlags}
\title{Filter out non-flagged rows on FlagOverTime Widget}
\usage{
FilterByFlags(dfResults, bCurrentlyFlagged = FALSE)
}
\arguments{
\item{dfResults}{`data.frame` Analysis results data.}

\item{bCurrentlyFlagged}{`logical` Include risk signals flagged in most recent snapshot?
Default: `FALSE`.}
}
\value{
A data frame containing the results with at least one flagged record over time for an
group's individual metric
}
\description{
Filter a results dataframe so that only metrics across all timepoints
that have at least one flag are kept
}
\examples{
reportingResults_flags <- FilterByFlags(gsm.core::reportingResults)

}
