% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-gt.R
\name{gsm_gt}
\alias{gsm_gt}
\title{Create a gt table object with gsm style}
\usage{
gsm_gt(data, ...)
}
\arguments{
\item{data}{A `gt_tbl` object.}

\item{...}{Additional arguments passed to [gt::gt()].}
}
\value{
An object of class `gt_tbl`.
}
\description{
@description
}
\details{
This is a wrapper to ensure that the user has gt installed and to apply
standardized styles.
}
\keyword{internal}
