% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeChartConfig.R
\name{MakeChartConfig}
\alias{MakeChartConfig}
\title{Make Chart Config}
\usage{
MakeChartConfig(lMetric, strChartFunction, ...)
}
\arguments{
\item{lMetric}{`list` Metric-specific metadata for use in charts and
reporting. Created by passing an `lWorkflow` object to [MakeMetric()] and
turing it into a list. Expected columns: `File`,`MetricID`, `Group`,
`Abbreviation`, `Metric`, `Numerator`, `Denominator`, `Model`, `Score`, and
`strThreshold`. For more details see the Data Model vignette:
`vignette("DataModel", package = "gsm.kri")`.}

\item{strChartFunction}{`character` Name of chart function.}

\item{...}{`any` Additional chart configuration settings.}
}
\value{
`list` Chart configuration.
}
\description{
Helper function to create chart configuration for a specific metric and chart type.
}
