% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeMetricTable.R
\name{MakeMetricTable}
\alias{MakeMetricTable}
\title{Generate a Summary data.frame for use in reports}
\usage{
MakeMetricTable(
  dfResults,
  dfGroups = NULL,
  strGroupLevel = c("Site", "Country", "Study"),
  strGroupDetailsParams = NULL,
  vFlags = c(-2, -1, 1, 2)
)
}
\arguments{
\item{dfResults}{`r gloss_param("dfResults")` `r gloss_extra("dfResults_filtered")`}

\item{dfGroups}{`data.frame` Group-level metadata dictionary. Created by
passing CTMS site and study data to [MakeLongMeta()]. Expected columns:
`GroupID`, `GroupLevel`, `Param`, `Value`.}

\item{strGroupLevel}{group level for the table}

\item{strGroupDetailsParams}{one or more parameters from dfGroups to be added
as columns in the table}

\item{vFlags}{`integer` List of flag values to include in output table.
Default: `c(-2, -1, 1, 2)`.}
}
\value{
A data.frame containing the summary table
}
\description{
`r lifecycle::badge("stable")`

Generate a summary table for a report by joining the provided results data
frame with the site-level metadata from dfGroups, and filter and arrange the
data based on provided conditions.
}
\examples{
# site-level report
MakeMetricTable(
  dfResults = gsm::reportingResults \%>\%
    dplyr::filter(.data$MetricID == "Analysis_kri0001") \%>\%
    FilterByLatestSnapshotDate(),
  dfGroups = gsm::reportingGroups
)

}
