% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeParamLabels.R
\name{MakeParamLabels}
\alias{MakeParamLabels}
\alias{MakeParamLabelsList}
\title{Create Labels for Parameters}
\usage{
MakeParamLabels(dfGroups, lParamLabels = NULL)

MakeParamLabelsList(chrParams, lParamLabels = NULL)
}
\arguments{
\item{chrParams}{A character vector of parameters, or a list that can be
coerced to a character vector.}
}
\value{
`dfGroups` with an added `Label` column, or a list of labeled
  parameters.
}
\description{
`r lifecycle::badge("stable")`

  Convert a vector of parameters to labels in Title Case. `MakeParamLabels`
  adds a `Labels` column to a `data.frame` that has a `Params` column (such
  as `dfGroups`), while `MakeParamLabelsList` returns just the list of named
  parameters.
}
\examples{
head(gsm::reportingGroups)
MakeParamLabels(head(gsm::reportingGroups))
MakeParamLabels(
  head(gsm::reportingGroups),
  list(ParticipantCount = "Number of Participants")
)
MakeParamLabelsList(head(gsm::reportingGroups$Param))

}
