% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeStudyInfo.R
\name{MakeStudyInfo}
\alias{MakeStudyInfo}
\title{Generate a study information data.frame for use in reports}
\usage{
MakeStudyInfo(dfGroups, lStudyLabels = NULL, lStudy = deprecated())
}
\arguments{
\item{dfGroups}{`data.frame` Group-level metadata dictionary. Created by
passing CTMS site and study data to [MakeLongMeta()]. Expected columns:
`GroupID`, `GroupLevel`, `Param`, `Value`.}

\item{lStudyLabels}{`list` A list containing study labels. Default is NULL.}

\item{lStudy}{`deprecated` Study information as a named list.}
}
\value{
A data.frame containing study metadata.
}
\description{
`r lifecycle::badge("stable")`

Generate a study info table summarizing study metadata.
}
\examples{
MakeStudyInfo(gsm::reportingGroups)
MakeStudyInfo(gsm::reportingGroups, list(SiteCount = "# Sites"))

}
