% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{Widget_BarChart}
\alias{Widget_BarChartOutput}
\alias{renderWidget_BarChart}
\alias{Widget_ScatterPlot}
\alias{Widget_ScatterPlotOutput}
\alias{renderWidget_ScatterPlot}
\alias{Widget_TimeSeries}
\alias{Widget_TimeSeriesOutput}
\alias{renderWidget_TimeSeries}
\alias{Widget_GroupOverview}
\alias{Widget_GroupOverviewOutput}
\alias{renderWidget_GroupOverview}
\alias{MakeChartConfig}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{gsm.vizr}{\code{\link[gsm.vizr:MakeChartConfig]{MakeChartConfig()}}, \code{\link[gsm.vizr:renderWidget_BarChart]{renderWidget_BarChart()}}, \code{\link[gsm.vizr:renderWidget_GroupOverview]{renderWidget_GroupOverview()}}, \code{\link[gsm.vizr:renderWidget_ScatterPlot]{renderWidget_ScatterPlot()}}, \code{\link[gsm.vizr:renderWidget_TimeSeries]{renderWidget_TimeSeries()}}, \code{\link[gsm.vizr:Widget_BarChart]{Widget_BarChart()}}, \code{\link[gsm.vizr:Widget_BarChartOutput]{Widget_BarChartOutput()}}, \code{\link[gsm.vizr:Widget_GroupOverview]{Widget_GroupOverview()}}, \code{\link[gsm.vizr:Widget_GroupOverviewOutput]{Widget_GroupOverviewOutput()}}, \code{\link[gsm.vizr:Widget_ScatterPlot]{Widget_ScatterPlot()}}, \code{\link[gsm.vizr:Widget_ScatterPlotOutput]{Widget_ScatterPlotOutput()}}, \code{\link[gsm.vizr:Widget_TimeSeries]{Widget_TimeSeries()}}, \code{\link[gsm.vizr:Widget_TimeSeriesOutput]{Widget_TimeSeriesOutput()}}}
}}

