% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_FlagChange.R
\name{Report_FlagChange}
\alias{Report_FlagChange}
\title{Report Flag Changes}
\usage{
Report_FlagChange(dfResults)
}
\arguments{
\item{dfResults}{A data.frame generated by gsm.reporting::CalculateChange().}
}
\value{
A character string containing an HTML unordered list of newly changed flags.
}
\description{
Create an HTML list of newly changed flags using the data generated by CalculateChange().
}
