% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_FormatFlag.R
\name{Report_FormatFlag}
\alias{Report_FormatFlag}
\title{KRI Directionality Logo.}
\usage{
Report_FormatFlag(flag_value, title = NULL)
}
\arguments{
\item{flag_value}{`numeric` Values between -2 and 2.}

\item{title}{`character` Data that will be passed to the <title> tag.}
}
\value{
A character vector of fontawesome icon SVGs.
}
\description{
`r lifecycle::badge("stable")`

Reporting utility function to create fontawesome directionality logos`
}
