% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_MetricCharts.R
\name{Report_MetricCharts}
\alias{Report_MetricCharts}
\title{Render charts for a given metric to markdown}
\usage{
Report_MetricCharts(lCharts, strMetricID = "")
}
\arguments{
\item{lCharts}{A list of charts for the selected metric.}

\item{strMetricID}{`character` Metric ID, wich is added as a class to each rendered chart to
uniquely identify it within the report.}
}
\value{
Markdown content with charts and a summary table for the metric
}
\description{
`r lifecycle::badge("stable")`

This function generates a markdown framework for charts
}
