% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_OverviewText.R
\name{Report_OverviewText}
\alias{Report_OverviewText}
\title{Create message describing study summary for Report}
\usage{
Report_OverviewText(lSetup, dfResults, lStudy)
}
\arguments{
\item{lSetup}{`list` that is produced by `Report_StudyInfo()`.}

\item{lStudy}{`list` contains study-level metadata.}
}
\description{
`r lifecycle::badge("stable")`
}
\keyword{internal}
