% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_Setup.R
\name{Report_Setup}
\alias{Report_Setup}
\title{Calculate needed values for report}
\usage{
Report_Setup(dfGroups = NULL, dfMetrics = NULL, dfResults = NULL)
}
\value{
`list` with the following elements:
- `GroupLevel` (character): The group level of the report.
- `SnapshotDate` (Date): The date of the snapshot.
- `lStudy` (list): Study-level metadata.
- `StudyID` (character): The study ID.
- `red_kris` (numeric): The number of red flags.
- `amber_kris` (numeric): The number of amber flags.
}
\description{
`r lifecycle::badge("stable")`
}
\keyword{internal}
