% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_StudyInfo.R
\name{Report_StudyInfo}
\alias{Report_StudyInfo}
\title{Report Study Information}
\usage{
Report_StudyInfo(
  dfGroups,
  lStudyLabels = NULL,
  strId = "study_table",
  tagHeader = htmltools::h2("Study Status"),
  lStudy = deprecated()
)
}
\arguments{
\item{lStudyLabels}{`list` A list containing study labels. Default is NULL.}

\item{strId}{`character` A string to identify the output table.}

\item{tagHeader}{`shiny.tag` An HTML tag or tags to use as a header for the
table.}

\item{lStudy}{`deprecated` Study information as a named list.}
}
\value{
A [htmltools::tagList()] to display a table of study information.
}
\description{
`r lifecycle::badge("stable")`

This function generates a table summarizing study metadata as an interactive
[gt::gt()] wrapped in HTML.
}
