% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiskScore.R
\name{RiskScore}
\alias{RiskScore}
\title{Calculate Risk Score}
\usage{
RiskScore(dfResults, dfMetricWeights = gsm.kri::metricWeights)
}
\arguments{
\item{dfResults}{`data.frame` The results data to calculate risk scores for.}

\item{dfMetricWeights}{`data.frame` Combinations of metric ID and flag value, each with a corresponding weight. Default: `gsm.kri::metricWeights`.}
}
\value{
`data.frame` `dfResults` with additional columns: `SnapshotMonth`, `RiskScore`, `RiskScoreMax`, `RiskScoreNormalized`, `nRed`, `nAmber`.
}
\description{
Calculates the risk score for each group in the provided results data frame.
}
\examples{
dfRiskScore <- RiskScore(gsm.core::reportingResults)

}
