% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-shared.R
\name{shared-params}
\alias{shared-params}
\title{Parameters used in multiple functions}
\arguments{
\item{bDebug}{`logical` Print debug messages? Default: `FALSE`.}

\item{dfBounds}{`data.frame` Set of predicted percentages/rates and upper-
and lower-bounds across the full range of sample sizes/total exposure
values for reporting. Created by passing `dfResults` and `dfMetrics` to
[MakeBounds()]. Expected columns: `Threshold`, `Denominator`, `Numerator`,
`Metric`, `MetricID`, `StudyID`, `SnapshotDate`.}

\item{dfGroups}{`data.frame` Group-level metadata dictionary. Created by
passing CTMS site and study data to [MakeLongMeta()]. Expected columns:
`GroupID`, `GroupLevel`, `Param`, `Value`.}

\item{dfInput}{`data.frame` Input data with one record per subject. Created
by passing Raw+ data into [Input_Rate()]. Expected columns: `GroupID`,
`GroupLevel`, `Numerator`, `Denominator` and/or columns specified in
`strCountCol` and `strGroupCol`.}

\item{dfMetrics}{`data.frame` Metric-specific metadata for use in charts and
reporting. Created by passing an `lWorkflow` object to [MakeMetric()].
Expected columns: `File`, `MetricID`, `Group`, `Abbreviation`, `Metric`,
`Numerator`, `Denominator`, `Model`, `Score`, and `Threshold`. For more
details see the Data Model vignette: `vignette("DataModel", package =
"gsm.core")`.}

\item{dfResults}{`data.frame` A stacked summary of analysis pipeline output.
Created by passing a list of results returned by [Summarize()] to
[BindResults()]. Expected columns: `GroupID`, `GroupLevel`, `Numerator`,
`Denominator`, `Metric`, `Score`, `Flag`, `MetricID`, `StudyID`,
`SnapshotDate`.}

\item{lMetric}{`list` Metric-specific metadata for use in charts and
reporting. Created by passing an `lWorkflow` object to [MakeMetric()] and
turing it into a list. Expected columns: `File`,`MetricID`, `Group`,
`Abbreviation`, `Metric`, `Numerator`, `Denominator`, `Model`, `Score`, and
`strThreshold`. For more details see the Data Model vignette:
`vignette("DataModel", package = "gsm.kri")`.}

\item{lParamLabels}{`list` Labels for parameters, with the parameters as
names, and the label as value.}

\item{strOutputLabel}{`character` Label describing the output object, which is assigned to the
`output_label` attribute of the output object and appears in the report generated by
[gsm.kri::Report_KRI()]..}

\item{vOutcomeOptions}{`character` Vector of outcome options for the y-axis dropdown.
Default: `c("Score", "Metric", "Numerator")`. Must include `strOutcome`.}
}
\description{
Reused parameter definitions are gathered here for easier usage.
}
\keyword{internal}
