% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiskScore.R
\name{SummarizeRiskScore}
\alias{SummarizeRiskScore}
\title{Summarize Risk Score}
\usage{
SummarizeRiskScore(dfRiskScoreTransposed, fnSummarize = mean)
}
\arguments{
\item{dfRiskScoreTransposed}{`data.frame` The transposed risk score data to be summarized.}

\item{fnSummarize}{`function` A function to summarize risk score. Default: `mean`.}
}
\value{
`data.frame` A summarized risk score data frame with aggregated values.
}
\description{
Summarizes group-level risk score across studies, returning the aggregate risk score.
}
\examples{
dfRiskScoreSummarized <- gsm.core::reportingResults \%>\%
    CalculateRiskScore(gsm.kri::metricWeights) \%>\%
    TransposeRiskScore() \%>\%
    SummarizeRiskScore()

}
