% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransposeRiskScore.R
\name{TransposeRiskScore}
\alias{TransposeRiskScore}
\title{Transpose Risk Score}
\usage{
TransposeRiskScore(dfResults, dfMetrics, strValuesFrom = "flag_icon")
}
\arguments{
\item{strValuesFrom}{`character` The column name from which to take values in the transposed data.}

\item{dfRiskScore}{`data.frame` The risk score data to be transposed.}
}
\value{
`data.frame` The transposed risk score data with `MetricID` as columns.
}
\description{
Transposes the risk score data frame so that each `MetricID` becomes a column, structuring the
data for reporting.
}
\examples{
dfRiskScoreTransposed <- gsm.core::reportingResults \%>\%
    CalculateRiskScore(gsm.kri::metricWeights) \%>\%
    TransposeRiskScore()

}
