% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_Metric.R
\name{Visualize_Metric}
\alias{Visualize_Metric}
\title{Visualize_Metric Function}
\usage{
Visualize_Metric(
  dfResults = dfResults,
  dfBounds = NULL,
  dfGroups = NULL,
  dfMetrics = NULL,
  strMetricID = NULL,
  strSnapshotDate = NULL,
  bDebug = FALSE
)
}
\arguments{
\item{strMetricID}{`character` MetricID to subset the data.}

\item{strSnapshotDate}{`character` Snapshot date to subset the data.}

\item{bDebug}{`logical` Display console in html viewer for debugging. Default is `FALSE`.}
}
\value{
A list containing the following charts:
- scatterPlot: A scatter plot using JavaScript.
- barChart: A bar chart using JavaScript with metric on the y-axis.
- timeSeries: A time series chart using JavaScript with score on the y-axis.
- metricTable: A table containing all
}
\description{
`r lifecycle::badge("stable")`

The function creates all available charts for a metric using the data provided
}
\examples{
lCharts <- Visualize_Metric(
  dfResults = reportingResults,
  dfBounds = reportingBounds,
  dfGroups = reportingGroups,
  dfMetrics = reportingMetrics,
  strMetricID = "Analysis_kri0001"
)

}
