% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_Score.R
\name{Visualize_Score}
\alias{Visualize_Score}
\title{Group-level visualization of scores.}
\usage{
Visualize_Score(
  dfResults,
  vThreshold = NULL,
  strType = "Metric",
  bFlagFilter = FALSE,
  strTitle = ""
)
}
\arguments{
\item{dfResults}{`data.frame` A stacked summary of analysis pipeline output.
Created by passing a list of results returned by [Summarize()] to
[BindResults()]. Expected columns: `GroupID`, `GroupLevel`, `Numerator`,
`Denominator`, `Metric`, `Score`, `Flag`, `MetricID`, `StudyID`,
`SnapshotDate`.}

\item{vThreshold}{`numeric` Threshold specification, a vector of length 2 that defaults to NULL.}

\item{strType}{`character` One of `"KRI"` or `"score"`.}

\item{bFlagFilter}{`logical` Filter out non-flagged groups? Default: `FALSE`}

\item{strTitle}{`character` Title of plot.}
}
\value{
group-level ggplot2 object.
}
\description{
`r lifecycle::badge("stable")`
}
\examples{
## Filter data to one metric and snapshot
reportingResults_filter <- reportingResults \%>\%
  dplyr::filter(MetricID == "Analysis_kri0001" & SnapshotDate == max(SnapshotDate))

Visualize_Score(dfResults = reportingResults_filter)

## Only show Flagged Groups
Visualize_Score(
  dfResults = reportingResults_filter,
  bFlagFilter = TRUE
)

## Custom Title
Visualize_Score(
  dfResults = reportingResults_filter,
  strTitle = "Custom Title"
)

}
