% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_CrossStudyRiskScore.R
\name{Widget_CrossStudyRiskScore-shiny}
\alias{Widget_CrossStudyRiskScore-shiny}
\alias{Widget_CrossStudyRiskScoreOutput}
\alias{renderWidget_CrossStudyRiskScore}
\title{Shiny bindings for Widget_CrossStudyRiskScore}
\usage{
Widget_CrossStudyRiskScoreOutput(outputId, width = "100\%", height = "600px")

renderWidget_CrossStudyRiskScore(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit}

\item{expr}{An expression that generates a Widget_CrossStudyRiskScore}

\item{env}{The environment in which to evaluate expr.}

\item{quoted}{Is expr a quoted expression?}
}
\description{
Shiny bindings for Widget_CrossStudyRiskScore
}
