% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_FlagOverTime.R
\name{Widget_FlagOverTime-shiny}
\alias{Widget_FlagOverTime-shiny}
\alias{Widget_FlagOverTimeOutput}
\alias{renderWidget_FlagOverTime}
\title{Shiny bindings for Widget_FlagOverTime}
\usage{
Widget_FlagOverTimeOutput(outputId, width = "100\%", height = "400px")

renderWidget_FlagOverTime(
  dfResults,
  dfMetrics,
  strGroupLevel = c("Site", "Study", "Country")
)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{dfResults}{`data.frame` A stacked summary of analysis pipeline output.
Created by passing a list of results returned by [Summarize()] to
[BindResults()]. Expected columns: `GroupID`, `GroupLevel`, `Numerator`,
`Denominator`, `Metric`, `Score`, `Flag`, `MetricID`, `StudyID`,
`SnapshotDate`.}

\item{dfMetrics}{`data.frame` Metric-specific metadata for use in charts and
reporting. Created by passing an `lWorkflow` object to [MakeMetric()].
Expected columns: `File`, `MetricID`, `Group`, `Abbreviation`, `Metric`,
`Numerator`, `Denominator`, `Model`, `Score`, and `Threshold`. For more
details see the Data Model vignette: `vignette("DataModel", package =
"gsm.core")`.}

\item{strGroupLevel}{`character` Value for the group level. Default: "Site".}
}
\description{
`r lifecycle::badge("stable")`

Output and render functions for using Widget_FlagOverTime within
Shiny applications and interactive Rmd documents.
}
