% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_FlagOverTime.R
\name{Widget_FlagOverTime}
\alias{Widget_FlagOverTime}
\title{Flag Over Time Widget}
\usage{
Widget_FlagOverTime(
  dfResults,
  dfMetrics,
  strGroupLevel = c("Site", "Study", "Country"),
  strFootnote = NULL,
  bExcludeEver = FALSE,
  bDebug = FALSE
)
}
\arguments{
\item{dfResults}{`data.frame` A stacked summary of analysis pipeline output.
Created by passing a list of results returned by [Summarize()] to
[BindResults()]. Expected columns: `GroupID`, `GroupLevel`, `Numerator`,
`Denominator`, `Metric`, `Score`, `Flag`, `MetricID`, `StudyID`,
`SnapshotDate`.}

\item{dfMetrics}{`data.frame` Metric-specific metadata for use in charts and
reporting. Created by passing an `lWorkflow` object to [MakeMetric()].
Expected columns: `File`, `MetricID`, `Group`, `Abbreviation`, `Metric`,
`Numerator`, `Denominator`, `Model`, `Score`, and `Threshold`. For more
details see the Data Model vignette: `vignette("DataModel", package =
"gsm.core")`.}

\item{strGroupLevel}{`character` Value for the group level. Default: "Site".}

\item{strFootnote}{`character` Text to insert for figure}

\item{bExcludeEver}{`logical` Exclude options in widget dropdown that include the string "ever".
Default: `FALSE`.}

\item{bDebug}{`logical` Print debug messages? Default: `FALSE`.}
}
\description{
`r lifecycle::badge("stable")`

A widget that generates a table of flags over time using
[Report_FlagOverTime()].
}
\examples{
# Include all risk signals, irrespective flag value.
Widget_FlagOverTime(
  dfResults = gsm.core::reportingResults,
  dfMetrics = gsm.core::reportingMetrics
)

# Include risk signals that were ever flagged.
Widget_FlagOverTime(
  dfResults = FilterByFlags(
    gsm.core::reportingResults
  ),
  dfMetrics = gsm.core::reportingMetrics
)

# Include risk signals that were only flagged in the most recent snapshot.
Widget_FlagOverTime(
  dfResults = FilterByFlags(
    gsm.core::reportingResults,
    bCurrentlyFlagged = TRUE
  ),
  dfMetrics = gsm.core::reportingMetrics,
  bExcludeEver = TRUE
)

}
