% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_PortfolioOverview.R
\name{Widget_PortfolioOverview-shiny}
\alias{Widget_PortfolioOverview-shiny}
\alias{Widget_PortfolioOverviewOutput}
\alias{renderWidget_PortfolioOverview}
\title{Shiny bindings for Widget_PortfolioOverview}
\usage{
Widget_PortfolioOverviewOutput(outputId, width = "100\%", height = "600px")

renderWidget_PortfolioOverview(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit}

\item{expr}{An expression that generates a Widget_PortfolioOverview}

\item{env}{The environment in which to evaluate expr.}

\item{quoted}{Is expr a quoted expression?}
}
\description{
Shiny bindings for Widget_PortfolioOverview
}
