% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_PrematureDeathBucketBar.R
\name{Widget_PrematureDeathBucketBar-shiny}
\alias{Widget_PrematureDeathBucketBar-shiny}
\alias{Widget_PrematureDeathBucketBarOutput}
\alias{renderWidget_PrematureDeathBucketBar}
\title{Shiny bindings for Widget_PrematureDeathBucketBar}
\usage{
Widget_PrematureDeathBucketBarOutput(
  outputId,
  width = "100\%",
  height = "400px"
)

renderWidget_PrematureDeathBucketBar(
  expr,
  env = parent.frame(),
  quoted = FALSE
)
}
\arguments{
\item{outputId, width, height, expr, env, quoted}{htmlwidgets Shiny binding params.}
}
\description{
Shiny bindings for Widget_PrematureDeathBucketBar
}
