% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_PrematureDeathReasonBar.R
\name{Widget_PrematureDeathReasonBar-shiny}
\alias{Widget_PrematureDeathReasonBar-shiny}
\alias{Widget_PrematureDeathReasonBarOutput}
\alias{renderWidget_PrematureDeathReasonBar}
\title{Shiny bindings for Widget_PrematureDeathReasonBar}
\usage{
Widget_PrematureDeathReasonBarOutput(
  outputId,
  width = "100\%",
  height = "400px"
)

renderWidget_PrematureDeathReasonBar(
  expr,
  env = parent.frame(),
  quoted = FALSE
)
}
\arguments{
\item{outputId, width, height, expr, env, quoted}{htmlwidgets Shiny binding params.}
}
\description{
Shiny bindings for Widget_PrematureDeathReasonBar
}
