% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_RecordDuplication.R
\name{Widget_RecordDuplication-shiny}
\alias{Widget_RecordDuplication-shiny}
\alias{Widget_RecordDuplicationOutput}
\alias{renderWidget_RecordDuplication}
\title{Shiny bindings for Widget_RecordDuplication}
\usage{
Widget_RecordDuplicationOutput(outputId, width = "100\%", height = "800px")

renderWidget_RecordDuplication(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit}

\item{expr}{An expression that generates a Widget_RecordDuplication}

\item{env}{The environment in which to evaluate expr.}

\item{quoted}{Is expr a quoted expression?}
}
\description{
Shiny bindings for Widget_RecordDuplication
}
