% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_RiskScore.R
\name{Widget_RiskScore-shiny}
\alias{Widget_RiskScore-shiny}
\alias{Widget_RiskScoreOutput}
\alias{renderWidget_RiskScore}
\title{Shiny bindings for Widget_RiskScore}
\usage{
Widget_RiskScoreOutput(outputId, width = "100\%", height = "400px")

renderWidget_RiskScore(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like '100%', '400px', 'auto') or a number, which will be coerced to a string and have 'px' appended.}

\item{expr}{An expression that generates a Widget_RiskScore}

\item{env}{The environment in which to evaluate expr.}

\item{quoted}{Is expr a quoted expression (with quote())? This is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using Widget_RiskScore within Shiny applications and interactive Rmd documents.
}
