% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_RiskScore.R, R/Widget_RiskScore.R
\name{Visualize_RiskScore}
\alias{Visualize_RiskScore}
\alias{Widget_RiskScore}
\title{Visualize Risk Score}
\usage{
Visualize_RiskScore(dfResults, dfMetrics, strGroupLevel = "Site")

Widget_RiskScore(
  dfResults,
  dfRiskScores,
  dfMetrics,
  strGroupLevel = "Site",
  bDebug = FALSE
)
}
\arguments{
\item{strGroupLevel}{`character` The group level to filter the risk score data. Default is 'Site'.}

\item{dfRiskScoreTransposed}{`data.frame` The summarized risk score data to be visualized.}
}
\description{
Generates a color-coded HTML table using htmlwidgets, mimicking the DT table but without DT dependencies.

`r lifecycle::badge("experimental")`

A widget that generates a color-coded HTML table of risk scores, mimicking the DT table but without DT dependencies.
}
\examples{
gsm.core::reportingResults \%>\%
    RiskScore(gsm.kri::metricWeights) \%>\%
    TransposeRiskScore() \%>\%
    Visualize_RiskScore()

gsm.core::reportingResults \%>\%
    CalculateRiskScore(gsm.kri::metricWeights) \%>\%
    TransposeRiskScore() \%>\%
    Widget_RiskScore()

}
