% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_ScatterPlot.R
\name{Widget_ScatterPlot}
\alias{Widget_ScatterPlot}
\title{Scatter Plot Widget}
\usage{
Widget_ScatterPlot(
  dfResults,
  lMetric = NULL,
  dfGroups = NULL,
  dfBounds = NULL,
  bAddGroupSelect = TRUE,
  strShinyGroupSelectID = "GroupID",
  bDebug = FALSE
)
}
\arguments{
\item{bAddGroupSelect}{`logical` Add a dropdown to highlight sites? Default: `TRUE`.}

\item{strShinyGroupSelectID}{`character` Element ID of group select in Shiny context. Default: `'GroupID'`.}
}
\description{
`r lifecycle::badge("stable")`

A widget that generates a scatter plot of group-level metric results, plotting the denominator
on the x-axis and the numerator on the y-axis.
}
\examples{
## Filter data to one metric and snapshot
reportingResults_filter <- reportingResults \%>\%
  dplyr::filter(MetricID == "Analysis_kri0001" & SnapshotDate == max(SnapshotDate))

reportingMetrics_filter <- reportingMetrics \%>\%
  dplyr::filter(MetricID == "Analysis_kri0001") \%>\%
  as.list()

reportingBounds_filter <- reportingBounds \%>\%
  dplyr::filter(MetricID == "Analysis_kri0001" & SnapshotDate == max(SnapshotDate))

Widget_ScatterPlot(
  dfResults = reportingResults_filter,
  lMetric = reportingMetrics_filter,
  dfGroups = reportingGroups,
  dfBounds = reportingBounds_filter
)

}
