% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateRiskScore.R
\name{CalculateRiskScore}
\alias{CalculateRiskScore}
\title{Calculate Risk Score}
\usage{
CalculateRiskScore(dfResults, dfMetricWeights = gsm.kri::metricWeights)
}
\arguments{
\item{dfResults}{`data.frame` A stacked summary of analysis pipeline output.
Created by passing a list of results returned by [Summarize()] to
[BindResults()]. Expected columns: `GroupID`, `GroupLevel`, `Numerator`,
`Denominator`, `Metric`, `Score`, `Flag`, `MetricID`, `StudyID`,
`SnapshotDate`.}

\item{dfMetricWeights}{`data.frame` Combinations of metric ID and flag value, each with a
corresponding weight.}
}
\value{
`data.frame` `dfResults` with the following additional columns:
- `SnapshotMonth`: The month of the snapshot, formatted as "YYYY-MM".
- `RiskScore`: The total risk score for the group.
- `RiskScoreMax`: The maximum possible risk score for the group.
- `RiskScoreNormalized`: The normalized risk score as a percentage of the maximum.
- `nRed`: The count of metrics flagged as red.
- `nAmber`: The count of metrics flagged as amber.
}
\description{
Calculates the risk score for each group in the provided results data frame.
}
\examples{
dfRiskScore <- CalculateRiskScore(gsm.core::reportingResults)

}
