% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateRiskScore.R
\name{CalculateRiskScore}
\alias{CalculateRiskScore}
\title{Calculate Risk Score}
\usage{
CalculateRiskScore(
  lAnalysis,
  dfMetricWeights = gsm.kri::metricWeights,
  dSnapshotDate = Sys.Date(),
  vThreshold = c(60, 30)
)
}
\arguments{
\item{lAnalysis}{`list` List of analysis outputs from the metrics calculated in the
`workflow/2_metrics` workflows}

\item{dfMetricWeights}{`data.frame` Combinations of metric ID and flag value, each with a
corresponding weight.}
}
\value{
`data.frame` `dfResults` with the following additional columns:
- `SnapshotMonth`: The month of the snapshot, formatted as "YYYY-MM".
- `RiskScore`: The total risk score for the group.
- `RiskScoreMax`: The maximum possible risk score for the group.
- `RiskScoreNormalized`: The normalized risk score as a percentage of the maximum.
- `nRed`: The count of metrics flagged as red.
- `nAmber`: The count of metrics flagged as amber.
}
\description{
Calculates the risk score for each group in the provided results data frame.
}
\examples{
lAnalysis <- list("Analysis_kri0001" = list(Analysis_Summary = gsm.core::analyticsSummary,
                                ID = "Analysis_kri0001"))
dfRiskScore <- CalculateRiskScore(lAnalysis)

}
