% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateRiskScore.R
\name{CalculateRiskScore}
\alias{CalculateRiskScore}
\title{Calculate Risk Score}
\usage{
CalculateRiskScore(lAnalysis)
}
\arguments{
\item{lAnalysis}{`list` List of analysis outputs from the metrics calculated in the
`workflow/2_metrics` workflows}

\item{dfMetricWeights}{`data.frame` Combinations of metric ID and flag value, each with a
corresponding weight.}

\item{dSnapshotDate}{`Date` The date of the snapshot. Default is the current date.}

\item{vThreshold}{`numeric` A vector of two numeric values representing the thresholds for flagging}
}
\value{
`data.frame` That has the same features as Analysis_Summary, but with the following additional columns:
- `SnapshotMonth`: The month of the snapshot in "YYYY-MM" format.
- `nAmber`: The count of metrics flagged as Amber.
- `nRed`: The count of metrics flagged as Red.
}
\description{
Calculates the risk score for each group in the provided results data frame.
}
\examples{
lAnalysis <- list("Analysis_kri0001" = list(Analysis_Summary = gsm.core::analyticsSummary,
                                ID = "Analysis_kri0001"))
dfRiskScore <- CalculateRiskScore(lAnalysis)

}
