% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateRiskScore.R
\name{CalculateRiskScore}
\alias{CalculateRiskScore}
\title{Calculate Risk Score}
\usage{
CalculateRiskScore(lAnalysis)
}
\arguments{
\item{lAnalysis}{`list` List of analysis outputs from the metrics calculated in the
`workflow/2_metrics` workflows}
}
\value{
`data.frame` That has the same features as Analysis_Summary.
}
\description{
Calculates the risk score for each group in the provided results data frame.
}
\examples{
analysisFlagged <- gsm.core::analyticsSummary \%>\%
  dplyr::mutate(Weight = dplyr::case_when(abs(Flag) == 1 ~ 2,
                                          abs(Flag) == 2 ~ 4,
                                          Flag == 0 ~ 0,
                                          TRUE ~ NA),
                WeightMax = 4)

lAnalysis <- list("Analysis_kri0001" = list(Analysis_Flagged = analysisFlagged,
                                ID = "Analysis_kri0001"))
dfRiskScore <- CalculateRiskScore(lAnalysis)

}
