% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsm.kri-package.R
\docType{package}
\name{gsm.kri-package}
\alias{gsm.kri}
\alias{gsm.kri-package}
\title{gsm.kri: Good Statistical Monitoring KRIs}
\description{
The Good Statistical Monitoring or 'gsm' suite of R packages provides a framework for statistical data monitoring. 'gsm.kri' provides workflows to generate metrics and functionality to visualize and report on these metrics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Gilead-BioStats/gsm.kri}
  \item \url{https://gilead-biostats.github.io/gsm.kri}
  \item Report bugs at \url{https://github.com/Gilead-BioStats/gsm.kri/issues}
}

}
\author{
\strong{Maintainer}: Jeremy Wildfire \email{jwildfire@gmail.com}

Authors:
\itemize{
  \item Laura Maxwell \email{lkmaxwell23@gmail.com}
}

Other contributors:
\itemize{
  \item Gilead Sciences [copyright holder]
}

}
\keyword{internal}
