% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeCharts.R
\name{MakeCharts}
\alias{MakeCharts}
\title{Helper function to create charts for multiple metrics}
\usage{
MakeCharts(dfResults, dfBounds, dfGroups, dfMetrics, bDebug = FALSE)
}
\value{
A list of charts for each metric.
}
\description{
`r lifecycle::badge("stable")`
}
