% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_PatientListing.R
\name{pd_PatientListing}
\alias{pd_PatientListing}
\title{Premature-death patient listing}
\usage{
pd_PatientListing(dfResults, dfDeath)
}
\arguments{
\item{dfResults}{`data.frame` Reporting results containing patient-level rows.}

\item{dfDeath}{`data.frame` Mapped death data keyed on `subjid`.}
}
\value{
A `DT::datatable` htmlwidget.
}
\description{
`r lifecycle::badge("experimental")`

`DT` table of flagged premature-death subjects.
}
