% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_KRI.R
\name{Report_KRI}
\alias{Report_KRI}
\title{Report_KRI function}
\usage{
Report_KRI(
  lCharts = NULL,
  dfResults = NULL,
  dfGroups = NULL,
  dfMetrics = NULL,
  strOutputDir = getwd(),
  strOutputFile = NULL
)
}
\arguments{
\item{lCharts}{A list of charts to include in the report.}

\item{strOutputDir}{The output directory path for the generated report. If not provided,
the report will be saved in the current working directory.}

\item{strOutputFile}{The output file name for the generated report. If not provided,
the report will be named based on the study ID, Group Level and Date.}
}
\value{
File path of the saved report html is returned invisibly. Save to object to view absolute output path.
}
\description{
`r lifecycle::badge("stable")`

This function generates a KRI report based on the provided inputs.
}
\examples{
\dontrun{
# Run site-level KRI report.
lChartsSite <- MakeCharts(
  dfResults = reportingResults,
  dfGroups = reportingGroups,
  dfMetrics = reportingMetrics,
  dfBounds = reportingBounds
)

strOutputFile <- "StandardSiteReport.html"
kri_report_path <- Report_KRI(
  lCharts = lChartsSite,
  dfResults = reportingResults,
  dfGroups = reportingGroups,
  dfMetrics = reportingMetrics,
  strOutputFile = strOutputFile
)

# Run country-level KRI report.
lChartsCountry <- MakeCharts(
  dfResults = reportingResults_country,
  dfGroups = reportingGroups_country,
  dfMetrics = reportingMetrics_country,
  dfBounds = reportingBounds_country
)

strOutputFile <- "StandardCountryReport.html"
kri_report_path <- Report_KRI(
  lCharts = lChartsCountry,
  dfResults = reportingResults_country,
  dfGroups = reportingGroups_country,
  dfMetrics = reportingMetrics_country,
  strOutputFile = strOutputFile
)
}

}
\keyword{KRI}
\keyword{report}
