% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_RiskScore.R
\name{Visualize_RiskScore}
\alias{Visualize_RiskScore}
\title{Visualize Risk Score}
\usage{
Visualize_RiskScore(dfRiskScoreTransposed, strGroupLevel = "Site")
}
\arguments{
\item{dfRiskScoreTransposed}{`data.frame` The summarized risk score data to be visualized.}

\item{strGroupLevel}{`character` The group level to filter the risk score data. Default is 'Site'.}
}
\description{
Visualizes the risk score data in a table format, allowing for easy comparison of risk scores across studies.
TODO: apply color palette to metric columns
}
\examples{
gsm.core::reportingResults \%>\%
    CalculateRiskScore(gsm.kri::metricWeights) \%>\%
    TransposeRiskScore() \%>\%
    Visualize_RiskScore()

}
