% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_RiskScore.R
\name{Visualize_RiskScore}
\alias{Visualize_RiskScore}
\title{Visualize Risk Score}
\usage{
Visualize_RiskScore(dfResults, dfMetrics, dfGroups, strGroupLevel = "Site")
}
\arguments{
\item{dfResults}{`data.frame` Analysis results from CalculateRiskScore}

\item{dfMetrics}{`data.frame` Metric metadata from gsm.core::reportingMetrics}

\item{dfGroups}{`data.frame` Group metadata from gsm.core::reportingGroups}

\item{strGroupLevel}{`character` The group level to filter the risk score
data. Default is 'Site'.}
}
\description{
Creates an interactive risk score widget for cross-study visualization.
}
\details{
For a working example see
[Cross-Study KRI Report](https://gilead-public.github.io/gsm.kri/examples/Example_CrossStudySRS.html).
}
