% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_BarChart.R
\name{Widget_BarChart}
\alias{Widget_BarChart}
\title{Bar Chart Widget}
\usage{
Widget_BarChart(
  dfResults,
  lMetric = NULL,
  dfGroups = NULL,
  vThreshold = NULL,
  strOutcome = "Score",
  bAddGroupSelect = TRUE,
  strShinyGroupSelectID = "GroupID",
  bDebug = FALSE
)
}
\arguments{
\item{vThreshold}{`numeric` Threshold values.}

\item{strOutcome}{`character` Outcome variable. Default: 'Score'.}

\item{bAddGroupSelect}{`logical` Add a dropdown to highlight sites? Default: `TRUE`.}

\item{strShinyGroupSelectID}{`character` Element ID of group select in Shiny context. Default: `'GroupID'`.}
}
\description{
`r lifecycle::badge("stable")`

A widget that generates a bar chart of group-level metric results, plotting groups on the x-axis
and the outcome (numerator, denominator, metric, or score) on the y-axis.
}
\examples{
## Filter data to one metric and snapshot
reportingResults_filter <- reportingResults \%>\%
  dplyr::filter(MetricID == "Analysis_kri0001" & SnapshotDate == max(SnapshotDate))

reportingMetrics_filter <- reportingMetrics \%>\%
  dplyr::filter(MetricID == "Analysis_kri0001") \%>\%
  as.list()

## Make chart
Widget_BarChart(
  dfResults = reportingResults_filter,
  dfGroups = reportingGroups,
  lMetric = reportingMetrics_filter,
  vThreshold = reportingMetrics_filter$Threshold
)

}
