% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_FlagOverTime.R
\name{Widget_FlagOverTime-shiny}
\alias{Widget_FlagOverTime-shiny}
\alias{Widget_FlagOverTimeOutput}
\alias{renderWidget_FlagOverTime}
\title{Shiny bindings for Widget_FlagOverTime}
\usage{
Widget_FlagOverTimeOutput(outputId, width = "100\%", height = "400px")

renderWidget_FlagOverTime(
  dfResults,
  dfMetrics,
  strGroupLevel = c("Site", "Study", "Country")
)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{strGroupLevel}{`character` Value for the group level. Default: "Site".}
}
\description{
`r lifecycle::badge("stable")`

Output and render functions for using Widget_FlagOverTime within
Shiny applications and interactive Rmd documents.
}
