% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_GroupOverview.R
\name{Widget_GroupOverview}
\alias{Widget_GroupOverview}
\title{Group Overview Widget}
\usage{
Widget_GroupOverview(
  dfResults,
  dfMetrics,
  dfGroups,
  strGroupLevel = NULL,
  strGroupSubset = "red",
  strGroupLabelKey = "InvestigatorLastName",
  bDebug = FALSE
)
}
\arguments{
\item{strGroupLevel}{`character` Value for the group level. Default: NULL and taken from `dfMetrics$GroupLevel` if available.}

\item{strGroupSubset}{`character` Subset of groups to include in the table. Default: 'red'. Options:
- 'all': All groups.
- 'red': Groups with 1+ red flags.
- 'red/amber': Groups with 1+ red/amber flag.
- 'amber': Groups with 1+ amber flag.}

\item{strGroupLabelKey}{`character` Value for the group label key. Default: 'InvestigatorLastName'.}
}
\description{
`r lifecycle::badge("stable")`

A widget that generates a group overview table of group-level metric results across one or more
metrics.
}
\examples{
# site-level report
Widget_GroupOverview(
  dfResults = FilterByLatestSnapshotDate(reportingResults),
  dfMetrics = reportingMetrics,
  dfGroups = reportingGroups
)

# filter site-level report to all flags
Widget_GroupOverview(
  dfResults = FilterByLatestSnapshotDate(reportingResults),
  dfMetrics = reportingMetrics,
  dfGroups = reportingGroups,
  strGroupSubset = "all"
)

# country-level report
reportingMetrics$GroupLevel <- "Country"
Widget_GroupOverview(
  dfResults = FilterByLatestSnapshotDate(reportingResults),
  dfMetrics = reportingMetrics,
  dfGroups = reportingGroups,
  strGroupLevel = "Country"
)

}
