% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_GroupOverview.R
\name{Widget_GroupOverview}
\alias{Widget_GroupOverview}
\title{Group Overview Widget}
\usage{
Widget_GroupOverview(
  dfResults,
  dfMetrics,
  dfGroups,
  strGroupLevel = NULL,
  strGroupSubset = "red",
  strGroupLabelKey = "InvestigatorLastName",
  bDebug = FALSE
)
}
\arguments{
\item{dfResults}{`data.frame` A stacked summary of analysis pipeline output.
Created by passing a list of results returned by [Summarize()] to
[BindResults()]. Expected columns: `GroupID`, `GroupLevel`, `Numerator`,
`Denominator`, `Metric`, `Score`, `Flag`, `MetricID`, `StudyID`,
`SnapshotDate`.}

\item{dfMetrics}{`data.frame` Metric-specific metadata for use in charts and
reporting. Created by passing an `lWorkflow` object to [MakeMetric()].
Expected columns: `File`, `MetricID`, `Group`, `Abbreviation`, `Metric`,
`Numerator`, `Denominator`, `Model`, `Score`, and `Threshold`. For more
details see the Data Model vignette: `vignette("DataModel", package =
"gsm")`.}

\item{dfGroups}{`data.frame` Group-level metadata dictionary. Created by
passing CTMS site and study data to [MakeLongMeta()]. Expected columns:
`GroupID`, `GroupLevel`, `Param`, `Value`.}

\item{strGroupLevel}{`character` Value for the group level. Default: NULL and taken from `dfMetrics$GroupLevel` if available.}

\item{strGroupSubset}{`character` Subset of groups to include in the table. Default: 'red'. Options:
- 'all': All groups.
- 'red': Groups with 1+ red flags.
- 'red/amber': Groups with 1+ red/amber flag.
- 'amber': Groups with 1+ amber flag.}

\item{strGroupLabelKey}{`character` Value for the group label key. Default: 'InvestigatorLastName'.}

\item{bDebug}{`logical` Print debug messages? Default: `FALSE`.}
}
\description{
`r lifecycle::badge("stable")`

A widget that generates a group overview table of group-level metric results across one or more
metrics.
}
\examples{
# site-level report
Widget_GroupOverview(
  dfResults = FilterByLatestSnapshotDate(reportingResults),
  dfMetrics = reportingMetrics,
  dfGroups = reportingGroups
)

# filter site-level report to all flags
Widget_GroupOverview(
  dfResults = FilterByLatestSnapshotDate(reportingResults),
  dfMetrics = reportingMetrics,
  dfGroups = reportingGroups,
  strGroupSubset = "all"
)

# country-level report
reportingMetrics$GroupLevel <- "Country"
Widget_GroupOverview(
  dfResults = FilterByLatestSnapshotDate(reportingResults),
  dfMetrics = reportingMetrics,
  dfGroups = reportingGroups,
  strGroupLevel = "Country"
)

}
