% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_TimeSeries.R
\name{Widget_TimeSeries}
\alias{Widget_TimeSeries}
\title{Time Series Widget}
\usage{
Widget_TimeSeries(
  dfResults,
  lMetric = NULL,
  dfGroups = NULL,
  vThreshold = NULL,
  strOutcome = "Score",
  bAddGroupSelect = TRUE,
  strShinyGroupSelectID = "GroupID",
  bDebug = FALSE
)
}
\arguments{
\item{vThreshold}{`numeric` Threshold value(s).}

\item{strOutcome}{`character` Outcome variable. Default: 'Score'.}

\item{bAddGroupSelect}{`logical` Add a dropdown to highlight sites? Default: `TRUE`.}

\item{strShinyGroupSelectID}{`character` Element ID of group select in Shiny context. Default: `'GroupID'`.}
}
\description{
`r lifecycle::badge("stable")`

A widget that generates a time series of group-level metric results over time, plotting snapshot
date on the x-axis and the outcome (numerator, denominator, metric, or score) on the y-axis.
}
\examples{
## Filter data to one metric
reportingResults_filter <- reportingResults \%>\%
  dplyr::filter(MetricID == "Analysis_kri0001")

reportingMetrics_filter <- reportingMetrics \%>\%
  dplyr::filter(MetricID == "Analysis_kri0001") \%>\%
  as.list()

Widget_TimeSeries(
  dfResults = reportingResults_filter,
  lMetric = reportingMetrics_filter,
  dfGroups = reportingGroups,
  vThreshold = reportingMetrics_filter$Threshold
)

}
